fix some type conversions

+++ src/buffer-static.c
@@ -25,7 +25,7 @@
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Initialise an empty buffer to hold 'nbytes'
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-static_buffer_t *init_buffer(int nbytes) {
+static_buffer_t *init_buffer(size_t nbytes) {
   static_buffer_t *buf = (static_buffer_t *)malloc(sizeof(static_buffer_t));
   if (buf == NULL) {
     error("init_buffer(): cannot malloc buffer");
@@ -65,7 +65,7 @@ void write_byte(R_outpstream_t stream, int c) {
 void write_bytes(R_outpstream_t stream, void *src, int length) {
   static_buffer_t *buf = (static_buffer_t *)stream->data;
   memcpy(buf->data + buf->pos, src, length);
-  buf->pos += length;
+  buf->pos += (size_t)length;
 }
 
 
@@ -86,5 +86,5 @@ int read_byte(R_inpstream_t stream) {
 void read_bytes(R_inpstream_t stream, void *dst, int length) {
   static_buffer_t *buf = (static_buffer_t *)stream->data;
   memcpy(dst, buf->data + buf->pos, length);
-  buf->pos += length;
+  buf->pos += (size_t)length;
 }
