+++ src/cctx.c
@@ -241,7 +241,7 @@ SEXP init_cctx_(SEXP level_, SEXP num_threads_, SEXP dict_) {
   if (!isNull(dict_)) {
     size_t status;
     if (TYPEOF(dict_) == RAWSXP) {
-      status = ZSTD_CCtx_loadDictionary(cctx, RAW(dict_), length(dict_));
+      status = ZSTD_CCtx_loadDictionary(cctx, RAW(dict_), (size_t)length(dict_));
     } else if (TYPEOF(dict_) == STRSXP) {
       const char *filename = CHAR(STRING_ELT(dict_, 0));
       size_t fsize;
